+++ hydra-tx/test/Hydra/Tx/Contract/Increment.hs
@@ -16,6 +16,7 @@ import Hydra.Contract.DepositError (DepositError (..))
 import Hydra.Contract.Error (toErrorCode)
 import Hydra.Contract.HeadError (HeadError (..))
 import Hydra.Contract.HeadState qualified as Head
+import Hydra.Contract.Util (UtilError (MintingOrBurningIsForbidden))
 import Hydra.Data.Party qualified as OnChain
 import Hydra.Ledger.Cardano.Time (slotNoFromUTCTime)
 import Hydra.Plutus.Orphans ()
@@ -37,10 +38,11 @@ import Hydra.Tx.Utils (adaOnly, verificationKeyToOnChainId)
 import PlutusLedgerApi.V2 qualified as Plutus
 import PlutusTx.Builtins (toBuiltin)
 import Test.Hydra.Tx.Fixture (aliceSk, bobSk, carolSk, slotLength, systemStart, testNetworkId, testPolicyId, testSeedInput)
-import Test.Hydra.Tx.Gen (genForParty, genScriptRegistry, genUTxOSized, genValue, genVerificationKey)
+import Test.Hydra.Tx.Gen (genForParty, genMintedOrBurnedValue, genScriptRegistry, genUTxOSized, genValue, genVerificationKey)
 import Test.Hydra.Tx.Mutation (
   Mutation (..),
   SomeMutation (..),
+  changeMintedTokens,
   modifyInlineDatum,
   replaceParties,
   replaceSnapshotVersion,
@@ -185,6 +187,8 @@ data IncrementMutation
   | -- | Add a second v_deposit input alongside an attacker-controlled
     -- output that redirects its value away from the head's continuation.
     IncrementAddExtraDepositInput
+  | -- | Minting or burning of tokens should not be possible in increment.
+    MutateTokenMintingOrBurning
   deriving stock (Generic, Show, Enum, Bounded)
 
 genIncrementMutation :: (Tx, UTxO) -> Gen SomeMutation
@@ -257,6 +261,8 @@ genIncrementMutation (tx, utxo) =
             [ AddInput extraIn extraDepositOut (Just $ toScriptData Claim)
             , AppendOutput attackerOut
             ]
+    , SomeMutation (pure $ toErrorCode MintingOrBurningIsForbidden) MutateTokenMintingOrBurning
+        <$> (changeMintedTokens tx =<< genMintedOrBurnedValue)
     ]
  where
   headTxOut = fromJust $ txOuts' tx !!? 0